plugins/enthusiast-file-parsers/pyproject.toml
[tool.poetry]
name = "enthusiast-file-parsers"
version = "1.0.0"
description = "A plugin for Enthusiast that provides a file parsers."
authors = ["Damian Sowiński <damian.sowinski@upsidelab.io>"]
readme = "README.md"

[tool.poetry.dependencies]
enthusiast-common = ">=1,<2"
pypdf = ">=6.0.0,<7.0.0"
pillow = ">=11.3.0,<12.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
